move MouseOverPopover to Shared
